Add CSS injector for testing and small CSS changes that we may need
I'm pretty sure that we will be doing a sub-theme but I haven't set one up yet so this is pretty much the easiest way to add some small CSS for the geshi filter  for testing right now. I also think it's a good idea to have this in case of emergencies to make small changes to CSS while waiting for a chance to push code.