Switch logs to file + tail instead of using /proc
Running containers on OpenShift implies that you need to be able to run
the container as a non-root user. Non-root users can not open stdout and
stderr file descriptors via /proc/self/fd/{1,2} unless they are part of
the tty group.

We can't add that group to the user in the Dockerfile
since the user/uid is determined at runtime by OpenShift.

We can't add the user to that group in the entrypoint since we'd have to
re-login and a non-root user can't change their own group membership
anyway

So for OpenShift, we will use log file + tail like are with the 'app'
container that will be communicating with the fpm container